+++ CHANGES.md
@@ -1,8 +1,12 @@
 # Version 0.7.5
 
+- Improved NeuroML conversion support
+
 - Make cfg.checkErrors = True by default, but set to False if using multiple cores or batch
 
 - Added methods to rename netParams dict keys in a nested format (needed for GUI)
+
+- Added analysis.plotSpikeStats() func to plot stats of cell rates, ISI CVs and synchronies 
 
 - Fixed issue to make convergence+divergence connections randomization more robust and efficient (issue #254)